+++ expo/gbe/forms/__init__.py
@@ -17,17 +17,11 @@ from remains import (
     CostumeBidSubmitForm,
     CostumeDetailsDraftForm,
     CostumeDetailsSubmitForm,
-    CueInfoForm,
-    LightingInfoForm,
     RehearsalSelectionForm,
     StageInfoForm,
     StageInfoSubmitForm,
     UserCreateForm,
     VendorBidForm,
-    VendorCueInfoForm,
-    VolunteerBidForm,
-    VolunteerInterestForm,
-    VolunteerOpportunityForm,
 )
 from contact_form import ContactForm
 from participant_form import ParticipantForm
@@ -49,3 +43,11 @@ from show_schedule_form import ShowScheduleForm
 from bid_evaluation_form import BidEvaluationForm
 from act_bid_evaluation_form import ActBidEvaluationForm
 from bid_state_change_form import BidStateChangeForm
+
+from volunteer_bid_form import VolunteerBidForm
+from volunteer_interest_form import VolunteerInterestForm
+from volunteer_opportunity_form import VolunteerOpportunityForm
+
+from cue_info_form import CueInfoForm
+from lighting_info_form import LightingInfoForm
+from vendor_cue_info_form import VendorCueInfoForm
